Add ChoreoLib, update PathplannerLib
I'm planning to likely use ChoreoLib for autonomous and PathplannerLib for on-the-fly paths in teleop, just because ChoreoLib syntax seems to be much nicer for building autos inside of the code. I don't think this should cause any issues, but I will likely fall back to PathplannerLib and just maybe make some custom wrappers if I have any issues.

Signed-off-by: Ashton A. <uhdbits@duck.com>