Get basic field constants finished
Basic poses for the robot (based on CAD dimensions) should now be correct. I still might work on doing this with some more dynamic creation of these poses in the code, for example, if our bumper size is slightly different than what is in CAD, we can just easily change the width without redoing all the calculations.

Signed-off-by: Ashton A. <uhdbits@duck.com>