Remove paths on the "top" side of the reef
I've realized that, just like we can flip the paths to work on the other alliance, we can also flip paths to work on the other side of the reef of our own alliance, so we can do more with less paths. This now means the primary reef scoring paths are done, now just have to work on starting position paths.

Signed-off-by: Ashton A. <uhdbits@duck.com>