Add ManagedStateMachine
Add the "ManagedStateMachine" class to provide the ability to easily have a state machine that does not update itself periodically through the command scheduler, and instead relies on an external "State Manager" (an ideal example is a full robot state machine) to update it's inputs. This provides the capability to update subsystems in a known order (the CommandScheduler does not order subsystems, so you can not rely on the periodic methods to be called in a reliable order), and all together, allowing better management and communication between subsystems.

Signed-off-by: Ashton A. <uhdbits@duck.com>