Remove startTransition(), move all to setState
startTransition() is only used once, and is private, so moving everything to the setState method just removes a unneeded method. Also add a new goalState variable to save the current goalState to prevent transition if the goalState is the same as the newState. This should prevent users from having to rewrite code for finding the transition state of transition states.

Signed-off-by: Ashton A. <uhdbits@duck.com>