Store final name in StateMachine, other naming changes
This commit stores a final string of the name in the StateMachine class, which should allow for very minor performance improvements, but more importantly, prevents logging values from changing if the subsystem name is changed through a "setName()" call. Also change lastTransitionTimestamp to lastStateChangeTimestamp, since transition() can be called on it's own, and it won't update the timestamp on it's own.

Signed-off-by: Ashton A. <uhdbits@duck.com>