Add alternate override fix to RobotManager
This does work, and won't set the managed state if an override is active, but I'm not sure if it's any faster or slower than just setting the managed state. I can't do accurate speed tests on the RIO, so I'm just gonna stick with the simpler solution for now until I can test.

Signed-off-by: Ashton A. <uhdbits@duck.com>